# rad-app/src/main/resources/rad-app/com/apihug/api.yaml
openapi: 3.0.1
info:
  title: rad
  description: apihug rapid application development template
  contact:
    name: developer@apihug.com
    url: https://github.com/apihug/
    email: developer@apihug.com
  version: 0.1.1-SNAPSHOT
...
paths:
...
  /api/customer/switch-tenant:
    post:
      tags:
      - customer
      summary: 切换租户
      description: 切换到指定租户，重新签发 Token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/SwitchTenantRequest"
      responses:
        "200":
          description: 切换到指定租户，重新签发 Token
          content:
            application/json:
              schema:
                type: object
                allOf:
                - $ref: "#/components/schemas/CommonResult"
                properties:
                  data:
                    $ref: "#/components/schemas/LoginResponse"
      x-hope-operation:
        requestClz: com.apihug.rad.api.customer.SwitchTenantRequest
        responseClz: com.apihug.rad.api.customer.LoginResponse
        questions:
        - 切换租户后返回什么？
        - 切换租户后权限如何变化？
        priority: MIDDLE
        authorization:
          lowLimitRiskyMode: LOGIN
  /api/customer/set-default/{tenantId}:
    post:
      tags:
      - customer
      summary: 设置默认租户
      description: 将指定租户设为客户的默认租户
      parameters:
      - name: tenantId
        in: path
        required: true
        allowEmptyValue: true
        schema:
          type: number
          format: int64
      x-hope-operation:
        questions:
        - 如何设置默认登录租户？
        - 不想切换租户？
        priority: MIDDLE
        authorization:
          lowLimitRiskyMode: LOGIN
...
  /api/customers/customers:
    post:
      tags:
      - customer
      summary: 创建客户
      description: 创建新客户
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "200":
          description: 创建新客户
          content:
            application/json:
              schema:
                type: object
                allOf:
                - $ref: "#/components/schemas/CommonResult"
                properties:
                  data:
                    $ref: "#/components/schemas/CustomerSummary"
      x-hope-operation:
        requestClz: com.apihug.rad.api.customer.CreateCustomerRequest
        responseClz: com.apihug.rad.api.customer.CustomerSummary
        questions:
        - 如何创建新客户？
        - 创建客户需要哪些信息？
        - 客户邮箱是否需要唯一？
        priority: HIGH
        authorization:
          rbac:
            authorities:
            - CUSTOMER_CREATE
  /api/customers/auth/forgot-password:
    post:
      tags:
      - auth
      summary: 找回密码
      description: 申请找回密码（发送重置邮件）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/ForgotPasswordRequest"
      x-hope-operation:
        requestClz: com.apihug.rad.api.customer.ForgotPasswordRequest
        questions:
        - 如何申请找回密码？
        - 重置邮件发送到哪里？
        priority: MIDDLE
        authorization:
          lowLimitRiskyMode: ANONYMOUS
  /api/customers/auth/reset-password:
    post:
      tags:
      - auth
      summary: 重置密码
      description: 重置密码（使用 token）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/ResetPasswordRequest"
      x-hope-operation:
        requestClz: com.apihug.rad.api.customer.ResetPasswordRequest
        questions:
        - 如何重置密码？
        - token 有效期多久？
        priority: HIGH
        authorization:
          lowLimitRiskyMode: ANONYMOUS
...
components:
  schemas:
    CommonError:
      type: object
      description: Standardized error response format used across the API.
      properties:
        code:
          type: integer
          format: int32
          description: Unique numeric error code for programmatic identification.
          example: 100002401
        title:
          type: string
          description: "Concise, uppercase error identifier (e.g., constant name)."
          example: USER_NAME_EMPTY
        description:
          type: string
          description: "Human-readable explanation of the error in the primary language\
            \ (e.g., English)."
          example: User name should not be empty
        description2:
          type: string
          description: "Human-readable explanation in a secondary language (e.g.,\
            \ Chinese)."
          example: 用户名不可以为空
        path:
          type: string
          description: "Dot-separated path to the invalid field (e.g., 'user.profile.email')."
          example: user.name
        field:
          type: string
          description: Name of the invalid field at the current nesting level.
          example: name
        rejectedValue:
          type: string
          description: The actual input value that was rejected during validation.
          example: aaron.blue
        constraint:
          type: string
          description: "Validation rule(s) that were violated (e.g., 'NOT_NULL', 'SIZE',\
            \ 'PATTERN'). Multiple rules may be pipe-separated."
          example: NOT_EMPTY|SIZE
        tips:
          type: string
          description: Actionable suggestion to resolve the error.
          example: "Price must be a positive number, e.g., 1000.12"
        domain:
          type: string
          description: "Business domain or module where the error originated (e.g.,\
            \ 'order', 'payment')."
          example: order
        phase:
          type: string
          description: The application layer or processing stage where the error was
            raised. Used for diagnostics and monitoring.
          enum:
          - CONTROLLER
          - SERVICE
          - DOMAIN
        severity:
          type: string
          description: The impact level of the error. Helps clients determine urgency
            and handling strategy.
          enum:
          - LOW
          - WARN
          - ERROR
          - FATAL
        httpStatus:
          type: string
          description: "The HTTP status code corresponding to this error, expressed\
            \ as the enum name (e.g., 'BAD_REQUEST', 'NOT_FOUND'). This typically\
            \ matches the actual HTTP response status sent by the server."
          enum:
          - UNDEFINE
          - BAD_REQUEST
          - UNAUTHORIZED
          - PAYMENT_REQUIRED
          - FORBIDDEN
          - NOT_FOUND
          - METHOD_NOT_ALLOWED
          - NOT_ACCEPTABLE
          - PROXY_AUTHENTICATION_REQUIRED
          - REQUEST_TIMEOUT
          - CONFLICT
          - GONE
          - LENGTH_REQUIRED
          - PRECONDITION_FAILED
          - PAYLOAD_TOO_LARGE
          - URI_TOO_LONG
          - UNSUPPORTED_MEDIA_TYPE
          - REQUESTED_RANGE_NOT_SATISFIABLE
          - EXPECTATION_FAILED
          - I_AM_A_TEAPOT
          - INSUFFICIENT_SPACE_ON_RESOURCE
          - METHOD_FAILURE
          - DESTINATION_LOCKED
          - UNPROCESSABLE_ENTITY
          - LOCKED
          - FAILED_DEPENDENCY
          - TOO_EARLY
          - UPGRADE_REQUIRED
          - PRECONDITION_REQUIRED
          - TOO_MANY_REQUESTS
          - REQUEST_HEADER_FIELDS_TOO_LARGE
          - UNAVAILABLE_FOR_LEGAL_REASONS
          - INTERNAL_SERVER_ERROR
          - NOT_IMPLEMENTED
          - BAD_GATEWAY
          - SERVICE_UNAVAILABLE
          - GATEWAY_TIMEOUT
          - HTTP_VERSION_NOT_SUPPORTED
          - VARIANT_ALSO_NEGOTIATES
          - INSUFFICIENT_STORAGE
          - LOOP_DETECTED
          - BANDWIDTH_LIMIT_EXCEEDED
          - NOT_EXTENDED
          - NETWORK_AUTHENTICATION_REQUIRED
        deprecated:
          type: boolean
          default: false
          description: Indicates whether this error type is deprecated and may be
            removed in future versions.
        properties:
          type: object
          additionalProperties: true
          description: Arbitrary key-value pairs for additional contextual information.
            These appear as top-level fields in JSON when serialized via Jackson's
            @JsonAnyGetter.
      title: (Common)Error
...
    CommonResult:
      type: object
      description: Standard API response wrapper
      properties:
        code:
          type: integer
          format: int32
          description: "Response status code: 0 indicates success; non-zero values\
            \ indicate errors."
          example: 0
        message:
          type: string
          description: Human-readable message providing additional context about the
            result.
          example: Operation succeeded
        errors:
          type: array
          description: List of error details (omitted if empty)
          items:
            $ref: "#/components/schemas/CommonError"
      title: (Common)Result Wrapper
...
    CustomerStatusEnum:
      type: string
      description: 客户状态枚举
      enum:
      - ACTIVE
      - INACTIVE
      - LOCKED
      - DISABLED
      x-hope-component:
        x-enum-extensions:
          ACTIVE:
            name: ACTIVE
            code: 1
            message: active
            message2: 活跃
          INACTIVE:
            name: INACTIVE
            code: 2
            message: inactive
            message2: 非活跃
          LOCKED:
            name: LOCKED
            code: 3
            message: locked
            message2: 锁定
          DISABLED:
            name: DISABLED
            code: 4
            message: disabled
            message2: 禁用
        compile:
          protoFile: com/apihug/rad/infra/customer/constant.proto
          name: CustomerStatusEnum
          locationLine: 8
          locationColumn: 1
        isEnum: true
        isError: false
        clz: com.apihug.rad.infra.customer.CustomerStatusEnum
...
    ForgotPasswordRequest:
      type: object
      description: 找回密码请求
      properties:
        email:
          type: string
          description: 邮箱
          maxLength: 100
          minLength: 1
          x-hope-field:
            mock:
              nature: EMAIL
            compile:
              locationLine: 539
              locationColumn: 3
              name: email
              type: java.lang.String
      title: ForgotPasswordRequest
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/customer/api.proto
          name: ForgotPasswordRequest
          locationLine: 531
          locationColumn: 1
        clz: com.apihug.rad.api.customer.ForgotPasswordRequest
...
    SwitchTenantRequest:
      type: object
      description: 切换租户请求
      properties:
        tenantId:
          type: number
          format: int64
          description: 目标租户 ID
          minimum: 1.0
          x-hope-field:
            compile:
              locationLine: 435
              locationColumn: 3
              name: tenant_id
              type: java.lang.Long
      title: SwitchTenantRequest
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/customer/api.proto
          name: SwitchTenantRequest
          locationLine: 427
          locationColumn: 1
        clz: com.apihug.rad.api.customer.SwitchTenantRequest
...
    ResetPasswordRequest:
      type: object
      description: 重置密码请求
      properties:
        token:
          type: string
          description: 重置 token
          maxLength: 255
          minLength: 1
          x-hope-field:
            mock:
              nature: GUID
            compile:
              locationLine: 556
              locationColumn: 3
              name: token
              type: java.lang.String
        password:
          type: string
          description: 新密码
          maxLength: 100
          minLength: 6
          x-hope-field:
            compile:
              locationLine: 564
              locationColumn: 3
              name: password
              type: java.lang.String
      title: ResetPasswordRequest
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/customer/api.proto
          name: ResetPasswordRequest
          locationLine: 548
          locationColumn: 1
        clz: com.apihug.rad.api.customer.ResetPasswordRequest
...
    TenantInfo:
      type: object
      description: 租户信息
      properties:
        id:
          type: number
          format: int64
          description: 租户 ID
          x-hope-field:
            compile:
              locationLine: 310
              locationColumn: 3
              name: id
              type: java.lang.Long
        tenantCode:
          type: string
          description: 租户代码
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 316
              locationColumn: 3
              name: tenant_code
              type: java.lang.String
        tenantName:
          type: string
          description: 租户名称
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 323
              locationColumn: 3
              name: tenant_name
              type: java.lang.String
        isDefault:
          type: boolean
          description: 是否默认租户
          x-hope-field:
            compile:
              locationLine: 330
              locationColumn: 3
              name: is_default
              type: java.lang.Boolean
        isPlatform:
          type: boolean
          description: 是否平台租户
          x-hope-field:
            compile:
              locationLine: 335
              locationColumn: 3
              name: is_platform
              type: java.lang.Boolean
      title: TenantInfo
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/customer/api.proto
          name: TenantInfo
          locationLine: 302
          locationColumn: 1
        clz: com.apihug.rad.api.customer.TenantInfo
...
    CustomerSummary:
      type: object
      description: 客户摘要信息
      properties:
        id:
          type: number
          format: int64
          description: 客户 ID
          minimum: 1.0
          x-hope-field:
            compile:
              locationLine: 496
              locationColumn: 3
              name: id
              type: java.lang.Long
        username:
          type: string
          description: 用户名
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 503
              locationColumn: 3
              name: username
              type: java.lang.String
        email:
          type: string
          description: 邮箱
          minLength: 1
          x-hope-field:
            mock:
              nature: EMAIL
            compile:
              locationLine: 510
              locationColumn: 3
              name: email
              type: java.lang.String
        status:
          type: object
          $ref: "#/components/schemas/CustomerStatusEnum"
          description: 客户状态
          enum:
          - ACTIVE
          - INACTIVE
          - LOCKED
          - DISABLED
          x-hope-field:
            compile:
              locationLine: 517
              locationColumn: 3
              name: status
              type: com.apihug.rad.infra.customer.CustomerStatusEnum
        tenantId:
          type: number
          format: int64
          description: 租户 ID
          x-hope-field:
            compile:
              locationLine: 523
              locationColumn: 3
              name: tenant_id
              type: java.lang.Long
      title: CustomerSummary
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/customer/api.proto
          name: CustomerSummary
          locationLine: 488
          locationColumn: 1
        clz: com.apihug.rad.api.customer.CustomerSummary
...
    LoginResponse:
      type: object
      description: 登录响应
      properties:
        accessToken:
          type: string
          description: 访问令牌
          minLength: 1
          x-hope-field:
            mock:
              nature: GUID
            compile:
              locationLine: 266
              locationColumn: 3
              name: access_token
              type: java.lang.String
        customerId:
          type: number
          format: int64
          description: 客户 ID
          minimum: 1.0
          x-hope-field:
            mock:
              numberRule:
                min: 1036
                max: 1000
            compile:
              locationLine: 273
              locationColumn: 3
              name: customer_id
              type: java.lang.Long
        username:
          type: string
          description: 用户名
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 281
              locationColumn: 3
              name: username
              type: java.lang.String
        needsTenantSelection:
          type: boolean
          description: 是否需要选择租户
          x-hope-field:
            compile:
              locationLine: 288
              locationColumn: 3
              name: needs_tenant_selection
              type: java.lang.Boolean
        tenants:
          type: array
          description: 客户的租户列表
          items:
            type: object
            $ref: "#/components/schemas/TenantInfo"
          x-hope-field:
            compile:
              locationLine: 293
              locationColumn: 3
              name: tenants
              type: com.apihug.rad.api.customer.TenantInfo
        defaultTenant:
          type: object
          $ref: "#/components/schemas/TenantInfo"
          description: 默认租户信息
          x-hope-field:
            compile:
              locationLine: 297
              locationColumn: 3
              name: default_tenant
              type: com.apihug.rad.api.customer.TenantInfo
      title: LoginResponse
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/customer/api.proto
          name: LoginResponse
          locationLine: 258
          locationColumn: 1
        clz: com.apihug.rad.api.customer.LoginResponse
...
    CreateCustomerRequest:
      type: object
      description: 创建客户请求
      properties:
        username:
          type: string
          description: 用户名
          maxLength: 50
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 451
              locationColumn: 3
              name: username
              type: java.lang.String
        password:
          type: string
          description: 密码
          maxLength: 100
          minLength: 6
          x-hope-field:
            compile:
              locationLine: 459
              locationColumn: 3
              name: password
              type: java.lang.String
        email:
          type: string
          description: 邮箱
          maxLength: 100
          minLength: 1
          x-hope-field:
            mock:
              nature: EMAIL
            compile:
              locationLine: 467
              locationColumn: 3
              name: email
              type: java.lang.String
        status:
          type: object
          $ref: "#/components/schemas/CustomerStatusEnum"
          description: 客户状态
          enum:
          - ACTIVE
          - INACTIVE
          - LOCKED
          - DISABLED
          x-hope-field:
            compile:
              locationLine: 475
              locationColumn: 3
              name: status
              type: com.apihug.rad.infra.customer.CustomerStatusEnum
        defaultTenantId:
          type: number
          format: int64
          description: 默认租户 ID
          minimum: 0.0
          x-hope-field:
            compile:
              locationLine: 481
              locationColumn: 3
              name: default_tenant_id
              type: java.lang.Long
      title: CreateCustomerRequest
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/customer/api.proto
          name: CreateCustomerRequest
          locationLine: 443
          locationColumn: 1
        clz: com.apihug.rad.api.customer.CreateCustomerRequest
...